Kotlin: convert vortexdbg-android linux/jni layer (wave C, round 3)
Convert 43 files: linux/ (AndroidElfLoader, LinuxModule, AndroidSyscall
Handler, ARM32SyscallHandler 2118 LOC, ARM64SyscallHandler 1608 LOC, init
functions, symbols), com/vortexdbg/AndroidEmulator, linux/android (16:
ArmLD/ArmLD64, emulators, builder, resolver, library files, hooks),
dvm/jni (15 Proxy* reflection classes), unpack, hook/xhook.
- api cross-cutting nullability: Loader.dlopen -> Module?, VirtualSymbol
  module nullable, AndroidEmulator.getDalvikVM -> VM?; emulator/ProxyDvmClass
  ctors internal; SocketIO AF_* constants Int (+ .toShort() at sockaddr
  family fields); ProxyReflected uses kotlin.Annotation; getName() !!.
- Verified: full-reactor BUILD SUCCESS (8 modules).

+++ vortexdbg-android/src/main/kotlin/com/vortexdbg/linux/file/DnsProxyDaemon.kt
@@ -154,7 +154,7 @@ internal class DnsProxyDaemon(private val sdk: Int) : LocalSocketIO.SocketHandle
             buffer.order(ByteOrder.BIG_ENDIAN)
             buffer.putInt(16) // ai_addrlen
             buffer.order(ByteOrder.LITTLE_ENDIAN)
-            buffer.putShort(SocketIO.AF_INET) // sin_family
+            buffer.putShort(SocketIO.AF_INET.toShort()) // sin_family
             buffer.putShort(java.lang.Short.reverseBytes(port)) // sin_port
             buffer.put(Arrays.copyOf(address.address, 4))
             buffer.put(ByteArray(8)) // __pad
@@ -169,7 +169,7 @@ internal class DnsProxyDaemon(private val sdk: Int) : LocalSocketIO.SocketHandle
             buffer.putInt(SocketIO.IPPROTO_TCP)
             buffer.putInt(16) // ai_addrlen
             buffer.order(ByteOrder.LITTLE_ENDIAN)
-            buffer.putShort(SocketIO.AF_INET) // sin_family
+            buffer.putShort(SocketIO.AF_INET.toShort()) // sin_family
             buffer.putShort(java.lang.Short.reverseBytes(port)) // sin_port
             buffer.put(Arrays.copyOf(address.address, 4))
             buffer.put(ByteArray(8)) // __pad
